[feat] docker-compose, express, db
docker-compose that starts the db, the api, the react app, and phpmyadmin

API is on port 4000
React app is on port 3000
PhpMyAdmin is ion port 8088

On first start if you got an error when doing `docker-compose up` (mysql2 can't connect to the db), just press CTRL+C and then docker-compose up again.

There is a button on the react app that calls the API just for some test. It's just the beginning, everything is really sketchy.